requirements.txt
asgiref==3.8.1
bleach==6.2.0
certifi==2025.1.31
charset-normalizer==3.4.1
cloudinary==1.44.0
diff-match-patch==20241021
Django==5.2
django-cloudinary-storage==0.3.0
django-cors-headers==4.7.0
django-import-export==4.3.7
django-summernote==0.8.20.0
djangorestframework==3.16.0
idna==3.10
pillow==11.2.1
python-dotenv==1.1.0
requests==2.32.3
six==1.17.0
sqlparse==0.5.3
tablib==3.8.0
urllib3==2.4.0
webencodings==0.5.1
gunicorn